Refactored out sticker window attachment
