+++ src/main/kotlin/io/unthrottled/doki/assets/AssetManager.kt
@@ -96,7 +96,7 @@ object AssetManager {
     assetPath: String,
     assetSource: String
   ): String = when (assetCategory) {
-    AssetCategory.STICKERS -> "$assetSource/${assetCategory.category}/vscode$assetPath"
+    AssetCategory.STICKERS -> "$assetSource/${assetCategory.category}/jetbrains$assetPath"
     else -> "$assetSource/${assetCategory.category}/$assetPath"
   }
