+++ src/main/kotlin/io/unthrottled/doki/settings/actors/StickerActor.kt
@@ -3,7 +3,7 @@ package io.unthrottled.doki.settings.actors
 import io.unthrottled.doki.config.ThemeConfig
 import io.unthrottled.doki.stickers.CurrentSticker
 import io.unthrottled.doki.stickers.StickerLevel
-import io.unthrottled.doki.stickers.StickerService
+import io.unthrottled.doki.stickers.StickerComponent
 import io.unthrottled.doki.themes.ThemeManager
 import io.unthrottled.doki.util.performWithAnimation
 
@@ -20,7 +20,7 @@ object StickerActor {
             CurrentSticker.DEFAULT
         }
         ThemeManager.instance.currentTheme.ifPresent {
-          StickerService.instance.activateForTheme(it)
+          StickerComponent.activateForTheme(it)
         }
       }
     }
@@ -32,11 +32,11 @@ object StickerActor {
         if (enabled) {
           ThemeConfig.instance.stickerLevel = StickerLevel.ON.name
           ThemeManager.instance.currentTheme.ifPresent {
-            StickerService.instance.activateForTheme(it)
+            StickerComponent.activateForTheme(it)
           }
         } else {
           ThemeConfig.instance.stickerLevel = StickerLevel.OFF.name
-          StickerService.instance.remove()
+          StickerComponent.remove()
         }
       }
     }
