+++ src/main/kotlin/io/unthrottled/doki/stickers/BackgroundWallpaperService.kt
@@ -4,14 +4,14 @@ import com.intellij.openapi.components.ServiceManager
 import io.unthrottled.doki.themes.DokiTheme
 import java.util.Optional
 
-interface StickerService {
+internal interface BackgroundWallpaperService {
 
   fun activateForTheme(dokiTheme: DokiTheme)
   fun remove()
 
   companion object {
-    val instance: StickerService
-      get() = ServiceManager.getService(StickerService::class.java)
+    val instance: BackgroundWallpaperService
+      get() = ServiceManager.getService(BackgroundWallpaperService::class.java)
   }
 
   fun checkForUpdates(dokiTheme: DokiTheme)
