+++ src/main/kotlin/io/unthrottled/doki/stickers/StickerComponent.kt
@@ -1,38 +1,54 @@
 package io.unthrottled.doki.stickers
 
+import com.intellij.ide.ui.LafManager
 import com.intellij.ide.ui.LafManagerListener
 import com.intellij.ide.ui.laf.LafManagerImpl
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.project.ProjectManagerListener
 import io.unthrottled.doki.config.ThemeConfig
+import io.unthrottled.doki.themes.DokiTheme
 import io.unthrottled.doki.themes.ThemeManager
 import io.unthrottled.doki.util.doOrElse
 import javax.swing.UIManager
 
-class StickerComponent : Disposable {
+class StickerComponent :
+  LafManagerListener,
+  Disposable {
   private val connection = ApplicationManager.getApplication().messageBus.connect()
 
   init {
+    StickerPanelService.instance.init()
     processLaf(LafManagerImpl.getInstance().currentLookAndFeel)
-    connection.subscribe(
-      LafManagerListener.TOPIC,
-      LafManagerListener {
-        processLaf(it.currentLookAndFeel)
-      }
-    )
+    connection.subscribe(LafManagerListener.TOPIC, this)
+  }
+
+  companion object {
+    fun activateForTheme(dokiTheme: DokiTheme) {
+      BackgroundWallpaperService.instance.activateForTheme(dokiTheme)
+      StickerPanelService.instance.activateForTheme(dokiTheme)
+    }
+
+    fun remove() {
+      BackgroundWallpaperService.instance.remove()
+      StickerPanelService.instance.remove()
+    }
+  }
+
+  override fun lookAndFeelChanged(source: LafManager) =
+    processLaf(source.currentLookAndFeel)
+
+  override fun dispose() {
+    connection.dispose()
   }
 
   private fun processLaf(currentLaf: UIManager.LookAndFeelInfo?) {
     ThemeManager.instance.processLaf(currentLaf)
       .filter { ThemeConfig.instance.currentStickerLevel == StickerLevel.ON }
       .doOrElse({
-        StickerService.instance.activateForTheme(it)
+        activateForTheme(it)
       }) {
-        StickerService.instance.remove()
+        remove()
       }
-  }
-
-  override fun dispose() {
-    connection.dispose()
   }
 }
